+++ packages/twopoint5d/src/stage/IProjection.ts
@@ -5,7 +5,7 @@ export interface IProjection {
   updateViewRect(width: number, height: number): void;
   getViewRect(): [width: number, height: number, pixelRatioHorizontal: number, pixelRatioVertical: number];
 
-  get projectionPlane(): ProjectionPlane;
+  get projectionPlane(): ProjectionPlane | undefined;
   getZoom(distanceToProjectionPlane: number): number;
 
   createCamera(): Camera;